requirements.txt
numpy>=1.21.0
pillow>=8.0.0
opencv-python>=4.5.0
ffmpeg-python>=0.2.0